+++ src/main/java/ladysnake/effective/EffectiveConfig.java
@@ -42,9 +42,6 @@ public class EffectiveConfig extends MidnightConfig {
 	@Entry
 	public static boolean allayTrails = true;
 
-	@Entry(min = 0, max = 100, isSlider = true)
-	public static int allayTwinkleDensity = 5;
-
 	@Entry
 	public static boolean goldenAllays = true;
 
@@ -80,6 +77,9 @@ public class EffectiveConfig extends MidnightConfig {
 
 	@Entry(min = 0, max = 100, isSlider = true)
 	public static float sculkParticleDensity = 100;
+
+	@Entry
+	public static boolean improvedSpectralArrows = true;
 
 
 	@Entry